backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.2
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.8.3 